fix merge, update Rider package

+++ Assets/Scripts/Components/SphereComponent.cs
@@ -4,6 +4,6 @@ namespace Tags
 {
     public struct SphereComponent : IComponentData
     {
-        public byte countElements;
+        public int countElements;
     }
 }
